+++ src/js/StairstepChart.js
@@ -1,6 +1,6 @@
 import Chart from 'chart.js';
 import * as Util from "./Util";
-import {cumulativeData, projectData} from "./main";
+import {cumulativeData, projectData, runningTotal} from "./main";
 import {LinearModel2d} from "../linear_model/Stats";
 
 // { data : Array [year, value],
@@ -25,6 +25,53 @@ export default function StairstepChart(config) {
             [2018, 2097]
         ];
     }());
+
+    function cumulativeData (data) {
+        let years = data.map(Util.first);
+        let values = data.map(Util.second);
+        return Util.zip(years, runningTotal(values));
+    }
+
+    // LinearModel -> Array [year, value] -> {
+    //   all : Array { date : Date, value : Number },
+    //   past : Array { date : Date, value : Number },
+    //   future : Array { date : Date, value : Number },
+    //   futureMinus : Array { date : Date, value : Number },
+    //   futurePlus : Array { date : Date, value : Number }
+    // }
+    // Will attempt to from 2016 to 2030.  Future data points are projected based on
+    // past data.
+    function projectData(linearmodel, origData) {
+        // Clone the data so we won't modify the
+        // original copy.
+        let data = Util.cloneArray(origData)
+            .map(Util.cloneArray);
+
+        let lastYear = Util.tail(data)[0];
+        // let linearmodel = new LinearModel2d(data);
+
+        for (let year = lastYear + 1; year <= 2030; year++) {
+            let projection = linearmodel.project_r_squared(year, year-lastYear+1);
+            data.push([year, Math.round(projection[1])]);
+        }
+
+        return data;
+    }
+
+    // Array a -> Getter a a Number Number -> Array Number
+    // Take an array of anything, and a function that gets numbers out of
+    // the array, and return an array of cumulative values
+     function runningTotal(data, lens) {
+        lens = lens || Util.identity;
+        let result = data.reduce(function(acc, value) {
+            let current = lens(value) + acc.total;
+            return {
+                data : acc.data.concat([current]),
+                total : current
+            }
+        }, { total : 0, data : [] });
+        return result.data;
+     }
 
     let boulderLinearModel = new LinearModel2d(boulderDataArray);
     let boulderData = projectData(
