+++ src/js/main.js
@@ -5,26 +5,10 @@ import 'vue-material/dist/theme/default.css';
 Vue.use(VueMaterial);
 
 import Tabs from './Tabs.js';
-import * as Util from './Util.js';
 import StairstepChart from './StairstepChart.js';
 import PieChart from './PieChart.js';
 
 import Main from '../App.vue';
-
-// Array a -> Getter a a Number Number -> Array Number
-// Take an array of anything, and a function that gets numbers out of
-// the array, and return an array of cumulative values  
-export function runningTotal(data, lens) {
-  lens = lens || Util.identity;
-  let result = data.reduce(function(acc, value) {
-    let current = lens(value) + acc.total;
-    return {
-      data : acc.data.concat([current]),
-      total : current
-    }
-  }, { total : 0, data : [] });
-  return result.data;  
-}
 
 // { selector : String, data :: Array [year, value] }
 export function run(options) {
@@ -80,39 +64,6 @@ let fortCollinsDataProcessor = function(apiRawResponse) {
         return [Number(key), Math.round(aggregatedByYear[key])];
     });
 };
-
-// Array [year, value] -> Array [year, value]
-export function cumulativeData (data) {
-  let years = data.map(Util.first);
-  let values = data.map(Util.second);
-  return Util.zip(years, runningTotal(values));
-}
-  
-// LinearModel -> Array [year, value] -> {
-//   all : Array { date : Date, value : Number },
-//   past : Array { date : Date, value : Number },
-//   future : Array { date : Date, value : Number },
-//   futureMinus : Array { date : Date, value : Number },
-//   futurePlus : Array { date : Date, value : Number }
-// }
-// Will attempt to from 2016 to 2030.  Future data points are projected based on
-// past data.
-export function projectData(linearmodel, origData) {
-  // Clone the data so we won't modify the
-  // original copy.
-  let data = Util.cloneArray(origData)
-                 .map(Util.cloneArray);
-
-  let lastYear = Util.tail(data)[0];
-  // let linearmodel = new LinearModel2d(data);
-    
-  for (let year = lastYear + 1; year <= 2030; year++) {
-    let projection = linearmodel.project_r_squared(year, year-lastYear+1);
-    data.push([year, Math.round(projection[1])]);
-  }
-
-  return data;
-}
 
 let xhr = new XMLHttpRequest();
 xhr.open('GET', 'https://opencity.fcgov.com/resource/ykei-s9zt.json');
